requirements.txt
selenium
requests
lxml
python-dotenv